requirements.txt
requests>=2.28
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
tenacity>=8.2.2
pandas>=2.0
